add BaskActivity

+++ app/src/main/java/com/happy/auction/entity/item/ItemGoods.java
@@ -13,7 +13,7 @@ public class ItemGoods extends BaseGoods {
     /**
      * 竞拍过期时间戳，单位：毫秒
      */
-    public int bid_expire_time;
+    public long bid_expire_time;
     /**
      * 状态:<br/>
      * 0:已结束，1：未开始，2：正在进行
